Now using another thread for HTTP request to overcome lag
Due to getInputStream being a blocking function and Minecraft being very single-thread based, HTTP requests were causing the game to freeze. To fix this I run getInputStream and rest of mod functionality from that point on on another to thread to ensure that the translated message is being correctly shown with text that is being translated and to allievate all game processes from being blocked.